+++ CHANGELOG.md
@@ -34,6 +34,7 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 * Refactoring:
   * `core/dom/image`
+  * `core/dom/resize-observer`
 
 ## v3.0.0-rc.141 (2021-02-05)